restructure files layout
