+++ transport_test.go
@@ -6,6 +6,6 @@ import (
 	test "github.com/libp2p/go-libp2p-testing/suites/mux"
 )
 
-func TestMultiplexTransport(t *testing.T) {
+func TestDefaultTransport(t *testing.T) {
 	test.SubtestAll(t, DefaultTransport)
 }
